Update issue 256
Acrivity のスキーマを調整しました。

git-svn-id: http://aipo.googlecode.com/svn/trunk@692 b7d4e767-949e-c8c6-5735-6c11b8a2f74f

+++ oms/src/main/java/com/aimluck/eip/cayenne/om/social/auto/_Activity.java
@@ -80,11 +80,11 @@ public class _Activity extends org.apache.cayenne.CayenneDataObject {
     }
     
     
-    public void setPriority(Integer priority) {
+    public void setPriority(Float priority) {
         writeProperty("priority", priority);
     }
-    public Integer getPriority() {
-        return (Integer)readProperty("priority");
+    public Float getPriority() {
+        return (Float)readProperty("priority");
     }
